requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-crontab>=0.7.1
celery>=5.3.0
redis>=4.5.0

